rename fuzz_data_loader to data_loader
